Refactor HttpSearch name (source of confusion with component)
